lint: fix

+++ packages/contracts-bedrock/contracts/EAS/EAS.sol
@@ -85,8 +85,7 @@ contract EAS is IEAS, Semver, EIP712Verifier {
     /**
      * @dev Creates a new EAS instance.
      */
-    constructor() Semver(1, 0, 0) EIP712Verifier("EAS", "1.0.0") {
-    }
+    constructor() Semver(1, 0, 0) EIP712Verifier("EAS", "1.0.0") {}
 
     /**
      * @inheritdoc IEAS
@@ -108,21 +107,30 @@ contract EAS is IEAS, Semver, EIP712Verifier {
     /**
      * @inheritdoc IEAS
      */
-    function attestByDelegation(
-        DelegatedAttestationRequest calldata delegatedRequest
-    ) external payable returns (bytes32) {
+    function attestByDelegation(DelegatedAttestationRequest calldata delegatedRequest)
+        external
+        payable
+        returns (bytes32)
+    {
         _verifyAttest(delegatedRequest);
 
         AttestationRequestData[] memory data = new AttestationRequestData[](1);
         data[0] = delegatedRequest.data;
 
-        return _attest(delegatedRequest.schema, data, delegatedRequest.attester, msg.value, true).uids[0];
+        return
+            _attest(delegatedRequest.schema, data, delegatedRequest.attester, msg.value, true).uids[
+                0
+            ];
     }
 
     /**
      * @inheritdoc IEAS
      */
-    function multiAttest(MultiAttestationRequest[] calldata multiRequests) external payable returns (bytes32[] memory) {
+    function multiAttest(MultiAttestationRequest[] calldata multiRequests)
+        external
+        payable
+        returns (bytes32[] memory)
+    {
         // Since a multi-attest call is going to make multiple attestations for multiple schemas, we'd need to collect
         // all the returned UIDs into a single list.
         bytes32[][] memory totalUids = new bytes32[][](multiRequests.length);
@@ -132,7 +140,7 @@ contract EAS is IEAS, Semver, EIP712Verifier {
         // from it to verify that there isn't any attempt to send too much ETH to resolvers. Please note that unless
         // some ETH was stuck in the contract by accident (which shouldn't happen in normal conditions), it won't be
         // possible to send too much ETH anyway.
-        uint availableValue = msg.value;
+        uint256 availableValue = msg.value;
 
         for (uint256 i = 0; i < multiRequests.length; i = uncheckedInc(i)) {
             // The last batch is handled slightly differently: if the total available ETH wasn't spent in full and there
@@ -182,7 +190,7 @@ contract EAS is IEAS, Semver, EIP712Verifier {
         // from it to verify that there isn't any attempt to send too much ETH to resolvers. Please note that unless
         // some ETH was stuck in the contract by accident (which shouldn't happen in normal conditions), it won't be
         // possible to send too much ETH anyway.
-        uint availableValue = msg.value;
+        uint256 availableValue = msg.value;
 
         for (uint256 i = 0; i < multiDelegatedRequests.length; i = uncheckedInc(i)) {
             // The last batch is handled slightly differently: if the total available ETH wasn't spent in full and there
@@ -193,7 +201,8 @@ contract EAS is IEAS, Semver, EIP712Verifier {
                 last = i == multiDelegatedRequests.length - 1;
             }
 
-            MultiDelegatedAttestationRequest calldata multiDelegatedRequest = multiDelegatedRequests[i];
+            MultiDelegatedAttestationRequest
+                calldata multiDelegatedRequest = multiDelegatedRequests[i];
             AttestationRequestData[] calldata data = multiDelegatedRequest.data;
 
             // Ensure that no inputs are missing.
@@ -249,7 +258,10 @@ contract EAS is IEAS, Semver, EIP712Verifier {
     /**
      * @inheritdoc IEAS
      */
-    function revokeByDelegation(DelegatedRevocationRequest calldata delegatedRequest) external payable {
+    function revokeByDelegation(DelegatedRevocationRequest calldata delegatedRequest)
+        external
+        payable
+    {
         _verifyRevoke(delegatedRequest);
 
         RevocationRequestData[] memory data = new RevocationRequestData[](1);
@@ -266,7 +278,7 @@ contract EAS is IEAS, Semver, EIP712Verifier {
         // from it to verify that there isn't any attempt to send too much ETH to resolvers. Please note that unless
         // some ETH was stuck in the contract by accident (which shouldn't happen in normal conditions), it won't be
         // possible to send too much ETH anyway.
-        uint availableValue = msg.value;
+        uint256 availableValue = msg.value;
 
         for (uint256 i = 0; i < multiRequests.length; i = uncheckedInc(i)) {
             // The last batch is handled slightly differently: if the total available ETH wasn't spent in full and there
@@ -280,7 +292,13 @@ contract EAS is IEAS, Semver, EIP712Verifier {
             MultiRevocationRequest calldata multiRequest = multiRequests[i];
 
             // Ensure to deduct the ETH that was forwarded to the resolver during the processing of this batch.
-            availableValue -= _revoke(multiRequest.schema, multiRequest.data, msg.sender, availableValue, last);
+            availableValue -= _revoke(
+                multiRequest.schema,
+                multiRequest.data,
+                msg.sender,
+                availableValue,
+                last
+            );
         }
     }
 
@@ -294,7 +312,7 @@ contract EAS is IEAS, Semver, EIP712Verifier {
         // from it to verify that there isn't any attempt to send too much ETH to resolvers. Please note that unless
         // some ETH was stuck in the contract by accident (which shouldn't happen in normal conditions), it won't be
         // possible to send too much ETH anyway.
-        uint availableValue = msg.value;
+        uint256 availableValue = msg.value;
 
         for (uint256 i = 0; i < multiDelegatedRequests.length; i = uncheckedInc(i)) {
             // The last batch is handled slightly differently: if the total available ETH wasn't spent in full and there
@@ -305,7 +323,9 @@ contract EAS is IEAS, Semver, EIP712Verifier {
                 last = i == multiDelegatedRequests.length - 1;
             }
 
-            MultiDelegatedRevocationRequest memory multiDelegatedRequest = multiDelegatedRequests[i];
+            MultiDelegatedRevocationRequest memory multiDelegatedRequest = multiDelegatedRequests[
+                i
+            ];
             RevocationRequestData[] memory data = multiDelegatedRequest.data;
 
             // Ensure that no inputs are missing.
@@ -438,7 +458,9 @@ contract EAS is IEAS, Semver, EIP712Verifier {
         res.uids = new bytes32[](length);
 
         // Ensure that we aren't attempting to attest to a non-existing schema.
-        SchemaRecord memory schemaRecord = ISchemaRegistry(Predeploys.SCHEMA_REGISTRY).getSchema(schema);
+        SchemaRecord memory schemaRecord = ISchemaRegistry(Predeploys.SCHEMA_REGISTRY).getSchema(
+            schema
+        );
         if (schemaRecord.uid == EMPTY_UID) {
             revert InvalidSchema();
         }
@@ -504,7 +526,14 @@ contract EAS is IEAS, Semver, EIP712Verifier {
             emit Attested(request.recipient, attester, uid, schema);
         }
 
-        res.usedValue = _resolveAttestations(schemaRecord, attestations, values, false, availableValue, last);
+        res.usedValue = _resolveAttestations(
+            schemaRecord,
+            attestations,
+            values,
+            false,
+            availableValue,
+            last
+        );
 
         return res;
     }
@@ -528,7 +557,9 @@ contract EAS is IEAS, Semver, EIP712Verifier {
         bool last
     ) private returns (uint256) {
         // Ensure that a non-existing schema ID wasn't passed by accident.
-        SchemaRecord memory schemaRecord = ISchemaRegistry(Predeploys.SCHEMA_REGISTRY).getSchema(schema);
+        SchemaRecord memory schemaRecord = ISchemaRegistry(Predeploys.SCHEMA_REGISTRY).getSchema(
+            schema
+        );
         if (schemaRecord.uid == EMPTY_UID) {
             revert InvalidSchema();
         }
@@ -660,7 +691,15 @@ contract EAS is IEAS, Semver, EIP712Verifier {
     ) private returns (uint256) {
         uint256 length = attestations.length;
         if (length == 1) {
-            return _resolveAttestation(schemaRecord, attestations[0], values[0], isRevocation, availableValue, last);
+            return
+                _resolveAttestation(
+                    schemaRecord,
+                    attestations[0],
+                    values[0],
+                    isRevocation,
+                    availableValue,
+                    last
+                );
         }
 
         ISchemaResolver resolver = schemaRecord.resolver;
@@ -773,7 +812,11 @@ contract EAS is IEAS, Semver, EIP712Verifier {
      * @param data The data to timestamp.
      * @param time The timestamp.
      */
-    function _revokeOffchain(address revoker, bytes32 data, uint64 time) private {
+    function _revokeOffchain(
+        address revoker,
+        bytes32 data,
+        uint64 time
+    ) private {
         mapping(bytes32 => uint64) storage revocations = _revocationsOffchain[revoker];
 
         if (revocations[data] != 0) {
@@ -801,7 +844,11 @@ contract EAS is IEAS, Semver, EIP712Verifier {
      *
      * @return A merged and flatten list of all the UIDs.
      */
-    function _mergeUIDs(bytes32[][] memory uidLists, uint256 uidsCount) private pure returns (bytes32[] memory) {
+    function _mergeUIDs(bytes32[][] memory uidLists, uint256 uidsCount)
+        private
+        pure
+        returns (bytes32[] memory)
+    {
         bytes32[] memory uids = new bytes32[](uidsCount);
 
         uint256 currentIndex = 0;
